+++ app-frontend/src/app/index.components.js
@@ -93,8 +93,8 @@ export default angular.module('index.components', [
     require('./components/filters/shapeFilter/shapeFilter.module.js').default.name,
     require('./components/filters/sliderFilter/sliderFilter.module.js').default.name,
     require('./components/filters/tagFilter/tagFilter.module.js').default.name,
+    require('./components/filters/ingestedFilter/ingestedFilter.module.js').default.name,
     ...exportComponents,
-
 
     // common components (no domain)
     require('./components/common/navBar/navBar.module.js').default.name,